backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
python-dotenv==1.0.1
loguru==0.7.2
pytest==8.3.2
httpx==0.27.2
ruff==0.6.9
black==24.8.0
bandit==1.7.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pydantic[email]==2.9.2
pytest-cov==5.0.0
mypy==1.11.2
bcrypt==4.1.3
